feat: Moved types to single file
- Attempt at keeping "how data translates" away from "what the data is".
- Some code refactoring
