+++ .editorconfig
@@ -1,0 +1,2 @@
+[*.{cs}]
+dotnet_diagnostic.IDE0130.severity = none
