+++ Assets/Scripts/RidersGame/Enums.cs
@@ -1,9 +1,9 @@
 namespace RidersCore.Data
 {
     public enum VehicleType
-        {
-            Skateboard,
-            Blades,
-            Bike
-        }
+    {
+        Skateboard,
+        Blades,
+        Bike
+    }
 }
