+++ Assets/Scripts/RidersGame/Environment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 761cc72111ca29f4f8c6ba206ef35cd7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
