+++ Assets/Scripts/RidersGame/Inputs/BaseInput.cs
@@ -1,27 +1,8 @@
 using UnityEngine;
+using RidersCore.Data;
 
 namespace RidersCore
 {
-    // All the relevant input data for the "actor" in the world
-    [System.Serializable]
-    public struct ActorInputData
-    {
-        public float Accelerate;
-        public float Brake;
-        // Alternatively
-        // public float VerticalInput;
-
-        public float TurnInput;
-
-        public bool Jump;
-        public float JumpHoldDuration;
-        public bool StuntA;
-        public bool StuntB;
-        public bool StuntC;
-        public bool Drift;
-        public bool BoostRam;
-    }
-
     public interface IInput
     {
         public ActorInputData GrabCurrentFrameInputs();
