--- Assets/Scripts/RidersGame/Inputs/InputFrameRecord.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0c8a7a750e8054e4e9bc3560fe451797
