--- Assets/Scripts/RidersGame/RacerInformation.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fd98d845641015840a62d5f6faab94ac
