+++ Assets/Scripts/RidersGame/Types.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 405705bdf3c6da440b41d9be1b9442fd
