+++ Assets/Scripts/RidersGame/VehicleSystems/Types.cs
@@ -1,0 +1,69 @@
+using UnityEngine;
+
+namespace RidersCore.Data
+{
+    [System.Serializable]
+    public struct VehicleStats
+    {
+        // figure out what stats we need
+        // currently just copy+pasting from microgame
+        [Header("Movement Settings")]
+        [Min(0.001f), Tooltip("Top speed attainable when moving forward.")]
+        public float TopSpeed;
+
+        [Tooltip("How quickly the kart reaches top speed.")]
+        public float Acceleration;
+
+        [Min(0.001f), Tooltip("Top speed attainable when moving backward.")]
+        public float ReverseSpeed;
+
+        [Tooltip("How quickly the kart reaches top speed, when moving backward.")]
+        public float ReverseAcceleration;
+
+        [Tooltip("How quickly the kart starts accelerating from 0. A higher number means it accelerates faster sooner.")]
+        [Range(0.2f, 1)]
+        public float AccelerationCurve;
+
+        [Tooltip("How quickly the kart slows down when the brake is applied.")]
+        public float Braking;
+
+        [Tooltip("How quickly the kart will reach a full stop when no inputs are made.")]
+        public float CoastingDrag;
+
+        [Range(0.0f, 1.0f)]
+        [Tooltip("The amount of side-to-side friction.")]
+        public float Grip;
+
+        [Tooltip("How tightly the kart can turn left or right.")]
+        public float Steer;
+
+        [Tooltip("Additional gravity for when the kart is in the air.")]
+        public float AddedGravity;
+
+        //additions
+        [Tooltip("Additional top speed when activating Boost")]
+        public float BoostTopSpeed;
+
+        [Tooltip("Additional acceleration immediately after Boost")]
+        public float BoostAccel;
+
+
+        // allow for stat adding for powerups.
+        public static VehicleStats operator +(VehicleStats a, VehicleStats b)
+        {
+            return new VehicleStats
+            {
+                Acceleration = a.Acceleration + b.Acceleration,
+                AccelerationCurve = a.AccelerationCurve + b.AccelerationCurve,
+                Braking = a.Braking + b.Braking,
+                CoastingDrag = a.CoastingDrag + b.CoastingDrag,
+                AddedGravity = a.AddedGravity + b.AddedGravity,
+                Grip = a.Grip + b.Grip,
+                ReverseAcceleration = a.ReverseAcceleration + b.ReverseAcceleration,
+                ReverseSpeed = a.ReverseSpeed + b.ReverseSpeed,
+                TopSpeed = a.TopSpeed + b.TopSpeed,
+                Steer = a.Steer + b.Steer,
+            };
+        }
+    }
+}
